create separate component modules
