+++ src/components/Searchbar.jsx
@@ -1,0 +1,47 @@
+import { motion } from 'framer-motion';
+
+const Search = ({ search ,handleSearch, children, isOpen, setIsOpen }) => {
+
+    return (
+    <motion.div 
+        className="searchbar flex flex-col items-center justify-center w-[40rem] absolute top-2 bg-white z-10"
+        data-isOpen={isOpen}
+        layout
+        transition={{ duration: .2}}
+        style={{borderRadius: '2rem', borderWidth: '1px'}}
+    >
+        <motion.input
+            type="text"
+            className="searchbarInput w-[96%] px-4 py-2 rounded-full text-sm text-gray-600 outline-none pl-7"
+            placeholder="Search country"
+            value={search}
+            onChange={handleSearch}
+            onFocus={() => setIsOpen(true)}
+            onBlur={() => 
+                setTimeout(() => {
+                    setIsOpen(false)
+                }, 100)
+            }
+            
+            layout
+        />
+        <motion.div layout className="searchResults w-full">
+            {isOpen && (
+            <>
+                <motion.div 
+                    layout
+                    className="p-2 w-full"
+                    initial={{ opacity: 0, y: -20 }}
+                    animate={{ opacity: 1, y: 0 }}
+                    exit={{ opacity: 0 }}
+                    transition={{ duration: .2, delay: .1 }}
+                >
+                    {children}
+                </motion.div>
+            </>
+            )}
+        </motion.div>
+    </motion.div>
+)};
+
+export default Search;
